requirements.txt
pandas>=1.4
numpy
google-play-scraper>=1.0.5
tqdm
python-dateutil
